:art: fix: split extensions to UIViewController & add assets and PlacesModel
