+++ PPangDelivery/PPangDelivery/Assets.xcassets/asian.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "asian.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
