+++ PPangDelivery/PPangDelivery/Assets.xcassets/barbeque.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "barbeque.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
