+++ PPangDelivery/PPangDelivery/Assets.xcassets/bunsik.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "bunsik.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
