+++ PPangDelivery/PPangDelivery/Assets.xcassets/cafe.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "cafe.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
