+++ PPangDelivery/PPangDelivery/Assets.xcassets/chicken2.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "chicken2.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
