+++ PPangDelivery/PPangDelivery/Assets.xcassets/chinese.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "chinese.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
