+++ PPangDelivery/PPangDelivery/Assets.xcassets/fastfood.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "fastfood.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
