+++ PPangDelivery/PPangDelivery/Assets.xcassets/jjim.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "jjim.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
