+++ PPangDelivery/PPangDelivery/Assets.xcassets/jokbo.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "jokbo.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
