+++ PPangDelivery/PPangDelivery/Assets.xcassets/korean.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "korean.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
