+++ PPangDelivery/PPangDelivery/Assets.xcassets/lunchbox.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "lunchbox.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
