+++ PPangDelivery/PPangDelivery/Assets.xcassets/midnight.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "midnight.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
