+++ PPangDelivery/PPangDelivery/Assets.xcassets/pizza.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "pizza.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
