+++ PPangDelivery/PPangDelivery/Assets.xcassets/sushi.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "sushi.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
