+++ PPangDelivery/PPangDelivery/Assets.xcassets/vegetable.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "vegetable.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
