+++ PPangDelivery/PPangDelivery/Assets.xcassets/western.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chicken.png",
+      "filename" : "western.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
