+++ PPangDelivery/PPangDelivery/Map/PopUpViewController.swift
@@ -12,47 +12,19 @@ import CoreLocation
 import NMapsMap
 import SnapKit
 import Then
+import DLRadioButton
 
 // MARK: - PPangTogetherPopup
 
 class PopUpViewController: UIViewController, UIPickerViewDelegate, UIPickerViewDataSource {
     
     private var messageText: String?
-    private var attributedMessageText: NSAttributedString?
     private var contentView: UIView?
     private let members: [String] = ["2","3","4","5","6","7","8"]
-    private let textViewPlaceHolder = "ex) 당당치킨 시켜먹어요"
-
-
-    func numberOfComponents(in pickerView: UIPickerView) -> Int {
-        return 1
-    }
-    
-    func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
-        return members.count
-    }
-        
-    func pickerView(_ pickerView: UIPickerView, viewForRow row: Int, forComponent component: Int, reusing view: UIView?) -> UIView {
-        pickerView.subviews.forEach {
-            $0.backgroundColor = .clear
-        }
-        
-        let pickerViewLabel = UILabel()
-        pickerViewLabel.text = members[row]
-        pickerViewLabel.textAlignment = .center
-        
-        return pickerViewLabel
-    }
-    
-    
+    let textViewPlaceHolder = "ex) 당당치킨 시켜먹어요"
     
     private lazy var containerView: UIView = {
         let view = UIView()
-        // popup animation test
-        //        UIView.animate(withDuration: 0.3, delay: 10, options: [], animations: {
-        //
-        //                    view.transform = CGAffineTransform(scaleX: 2.0, y: 2.0)
-        //                }, completion: nil)
         view.backgroundColor = .white
         view.layer.cornerRadius = 20
         view.transform = CGAffineTransform(scaleX: 2.0, y: 2.0)
@@ -107,7 +79,7 @@ class PopUpViewController: UIViewController, UIPickerViewDelegate, UIPickerViewD
         $0.distribution = .fill
     }
     
-    lazy var numberOfMemberStackView = UIStackView().then {
+    private lazy var numberOfMemberStackView = UIStackView().then {
         $0.spacing = 5.0
         $0.distribution = .fill
         $0.snp.makeConstraints { make in
@@ -115,22 +87,19 @@ class PopUpViewController: UIViewController, UIPickerViewDelegate, UIPickerViewD
         }
     }
         
-    private lazy var userInputText = UITextView().then {
+    lazy var userInputText = UITextView().then {
         // text
         $0.text = textViewPlaceHolder
         $0.textColor = .lightGray
         $0.font = .systemFont(ofSize: 15.0)
         $0.textAlignment = NSTextAlignment.justified
 
-        // layer
         $0.layer.borderWidth = 1
         $0.layer.borderColor = UIColor.gray.cgColor
         
-        // etc.
         $0.returnKeyType = .done
         $0.isEditable = true
         
-        // Inset
         $0.textContainerInset = UIEdgeInsets(top: 10, left: 10, bottom: 10, right: 10);
         $0.contentInset = .zero
         $0.scrollIndicatorInsets = .zero
@@ -139,26 +108,22 @@ class PopUpViewController: UIViewController, UIPickerViewDelegate, UIPickerViewD
     }
     
     private lazy var messageLabel: UILabel? = {
-        guard messageText != nil || attributedMessageText != nil else { return nil }
+        guard messageText != nil else {return nil}
         
         let label = UILabel()
-        label.text = ("빵하실 위치 :  "+messageText!)
+        label.text = ("빵하실 위치 :  " + messageText!)
         label.font = .systemFont(ofSize: 10.0)
         label.textColor = .gray
+        label.textAlignment = .center
         label.numberOfLines = 0
         
-        if let attributedMessageText = attributedMessageText {
-            label.attributedText = attributedMessageText
-        }
         return label
     }()
     
-    convenience init(messageText: String? = nil,
-                     attributedMessageText: NSAttributedString? = nil) {
+    convenience init(messageText: String? = nil) {
         self.init()
         
         self.messageText = messageText
-        self.attributedMessageText = attributedMessageText
         modalPresentationStyle = .overFullScreen
     }
     
@@ -226,6 +191,8 @@ class PopUpViewController: UIViewController, UIPickerViewDelegate, UIPickerViewD
     
     public func getDueTime(){
         let label = UILabel()
+//        let dateLabel = UILabel()
+//        let dateFormatter = DateFormatter()
         let datePicker = UIDatePicker()
         
         label.text = "몇시까지 모집할까요?"
@@ -235,7 +202,18 @@ class PopUpViewController: UIViewController, UIPickerViewDelegate, UIPickerViewD
         datePicker.locale = Locale(identifier: "ko-KR")
         datePicker.timeZone = .autoupdatingCurrent
         datePicker.minimumDate = NSDate() as Date
+        
+//        view.backgroundColor = .clear
+//        view.addSubview(datePicker)
+//        view.snp.makeConstraints { make in
+//            make.right =
+//        }
+        
+//        dateFormatter.dateFormat = "HH:mm"
+//        dateLabel.backgroundColor = .red
+//        dateLabel.text = dateFormatter.string(from: datePicker.date)
 
+        
         dueTime.addArrangedSubview(label)
         dueTime.addArrangedSubview(datePicker)
     }
@@ -255,7 +233,7 @@ class PopUpViewController: UIViewController, UIPickerViewDelegate, UIPickerViewD
 
     }
     public func makeFoodStack() {
-        let stackFoodList = ["족발/보쌈", "찜/탕/찌개", "돈까스/회/일식", "피자", "치킨", "고기/구이", "야식", "양식", "중식", "아시안", "백반/죽/국수", "도시락", "분식", "카페/디저트", "패스트푸드", "채식"]
+        let stackFoodList = ["jokbo", "jjim", "sushi", "pizza", "chicken", "barbeque", "midnight", "western", "chinese", "asian", "korean", "lunchbox", "bunsik", "cafe", "fastfood", "vegetable"]
         let foodTitleLabel = UILabel()
         foodTitleLabel.text = "메뉴 종류를 선택 해 주세요"
         foodTitleLabel.textAlignment = .center
@@ -263,31 +241,35 @@ class PopUpViewController: UIViewController, UIPickerViewDelegate, UIPickerViewD
         foodTitleLabel.textColor = .black
         foodTitleLabel.numberOfLines = 0
         foodTitleLabel.heightAnchor.constraint(equalToConstant: 25.0).isActive = true
+        let RadioButtonGroup = DLRadioButton()
+        RadioButtonGroup.layer.opacity = 0
+        RadioButtonGroup.heightAnchor.constraint(equalToConstant: 1.0).isActive = true
         
         for (index, foodText) in stackFoodList.enumerated() {
-            let button = UIButton()
-            button.setImage(UIImage(named: "chicken"), for: .normal)
-            button.imageView?.contentMode = .scaleAspectFit
-            button.snp.makeConstraints { make in
-                //                make.width.equalTo(containerView.snp.width).dividedBy(4)
-                make.height.width.equalTo(66)
+            let foodButton = DLRadioButton()
+            foodButton.icon = UIImage(named: foodText) ?? RadioButtonGroup.icon
+//            foodButton.iconSelected = UIImage(named: "test") ?? RadioButtonGroup.iconSelected
+//            button.setImage(UIImage(named: "chicken"), for: .normal)
+//            button.imageView?.contentMode = .scaleAspectFit
+            foodButton.snp.makeConstraints { make in
+                make.height.width.equalTo(50)
             }
+            RadioButtonGroup.otherButtons.append(foodButton)
             switch (index % 4) {
-            case 0: v1FoodStackView.addArrangedSubview(button)
-            case 1: v2FoodStackView.addArrangedSubview(button)
-            case 2: v3FoodStackView.addArrangedSubview(button)
-            case 3: v4FoodStackView.addArrangedSubview(button)
+            case 0: v1FoodStackView.addArrangedSubview(foodButton)
+            case 1: v2FoodStackView.addArrangedSubview(foodButton)
+            case 2: v3FoodStackView.addArrangedSubview(foodButton)
+            case 3: v4FoodStackView.addArrangedSubview(foodButton)
             default:
                 break
             }
         }
         foodStackView.addArrangedSubview(foodTitleLabel)
+        foodStackView.addArrangedSubview(RadioButtonGroup)
         foodStackView.addArrangedSubview(v1FoodStackView)
         foodStackView.addArrangedSubview(v2FoodStackView)
         foodStackView.addArrangedSubview(v3FoodStackView)
         foodStackView.addArrangedSubview(v4FoodStackView)
-        //button action code
-        //view.itemButton.addTarget(self, action: #selector(buttonTap), for: .touchUpInside)
     }
     
     private func setupViews() {
@@ -332,119 +314,24 @@ class PopUpViewController: UIViewController, UIPickerViewDelegate, UIPickerViewD
             make.width.equalTo(containerStackView.snp.width)
         }
     }
-}
-
-
-// MARK: - Extension
-
-extension PopUpViewController: UITextViewDelegate {
-    public func textViewDidBeginEditing(_ textView: UITextView) {
-        if userInputText.text == textViewPlaceHolder {
-            userInputText.text = nil
-            userInputText.textColor = .black
-        }
-    }
-    func PopUpViewController(_ textView: UITextView) {
-        if userInputText.text.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty {
-            userInputText.text = textViewPlaceHolder
-            userInputText.textColor = .lightGray
-        }
-    }
-}
-
-extension UIColor {
-    func image(_ size: CGSize = CGSize(width: 1, height: 1)) -> UIImage {
-        return UIGraphicsImageRenderer(size: size).image { rendererContext in
-            self.setFill()
-            rendererContext.fill(CGRect(origin: .zero, size: size))
-        }
-    }
-}
-
-extension UIControl {
-    public typealias UIControlTargetClosure = (UIControl) -> ()
     
-    private class UIControlClosureWrapper: NSObject {
-        let closure: UIControlTargetClosure
-        init(_ closure: @escaping UIControlTargetClosure) {
-            self.closure = closure
-        }
+    func numberOfComponents(in pickerView: UIPickerView) -> Int {
+        return 1
     }
     
-    private struct AssociatedKeys {
-        static var targetClosure = "targetClosure"
+    func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
+        return members.count
     }
-    
-    private var targetClosure: UIControlTargetClosure? {
-        get {
-            guard let closureWrapper = objc_getAssociatedObject(self, &AssociatedKeys.targetClosure) as? UIControlClosureWrapper else { return nil }
-            return closureWrapper.closure
-            
-        } set(newValue) {
-            guard let newValue = newValue else { return }
-            objc_setAssociatedObject(self, &AssociatedKeys.targetClosure, UIControlClosureWrapper(newValue),
-                                     objc_AssociationPolicy.OBJC_ASSOCIATION_RETAIN_NONATOMIC)
-        }
-    }
-    
-    @objc func closureAction() {
-        guard let targetClosure = targetClosure else { return }
-        targetClosure(self)
-    }
-    
-    public func addAction(for event: UIControl.Event, closure: @escaping UIControlTargetClosure) {
-        targetClosure = closure
-        addTarget(self, action: #selector(UIControl.closureAction), for: event)
-    }
-    
-}
-
-extension UIViewController {
-    func showPopUp(message: String? = nil,
-                   attributedMessage: NSAttributedString? = nil,
-                   leftActionTitle: String? = "취소",
-                   rightActionTitle: String = "확인",
-                   leftActionCompletion: (() -> Void)? = nil,
-                   rightActionCompletion: (() -> Void)? = nil) {
-        let popUpViewController = PopUpViewController(messageText: message,
-                                                      attributedMessageText: attributedMessage)
-        showPopUp(popUpViewController: popUpViewController,
-                  leftActionTitle: leftActionTitle,
-                  rightActionTitle: rightActionTitle,
-                  leftActionCompletion: leftActionCompletion,
-                  rightActionCompletion: rightActionCompletion)
-    }
-    
-    func showPopUp(contentView: UIView,
-                   leftActionTitle: String? = "취소",
-                   rightActionTitle: String = "확인",
-                   leftActionCompletion: (() -> Void)? = nil,
-                   rightActionCompletion: (() -> Void)? = nil) {
-        let popUpViewController = PopUpViewController(contentView: contentView)
         
-        showPopUp(popUpViewController: popUpViewController,
-                  leftActionTitle: leftActionTitle,
-                  rightActionTitle: rightActionTitle,
-                  leftActionCompletion: leftActionCompletion,
-                  rightActionCompletion: rightActionCompletion)
-    }
-    
-    private func showPopUp(popUpViewController: PopUpViewController,
-                           leftActionTitle: String?,
-                           rightActionTitle: String,
-                           leftActionCompletion: (() -> Void)?,
-                           rightActionCompletion: (() -> Void)?) {
-        popUpViewController.addActionToButton(title: leftActionTitle,
-                                              titleColor: .systemGray,
-                                              backgroundColor: .secondarySystemBackground) {
-            popUpViewController.dismiss(animated: false, completion: leftActionCompletion)
+    func pickerView(_ pickerView: UIPickerView, viewForRow row: Int, forComponent component: Int, reusing view: UIView?) -> UIView {
+        pickerView.subviews.forEach {
+            $0.backgroundColor = .clear
         }
+        let pickerViewLabel = UILabel()
+        pickerViewLabel.text = members[row]
+        pickerViewLabel.textAlignment = .center
         
-        popUpViewController.addActionToButton(title: rightActionTitle,
-                                              titleColor: .white,
-                                              backgroundColor: .blue) {
-            popUpViewController.dismiss(animated: false, completion: rightActionCompletion)
-        }
-        present(popUpViewController, animated: false, completion: nil)
+        return pickerViewLabel
     }
+
 }
